Crate sentry_failure

Adds support for capturing Sentry errors from failure types.

Failure errors and Fail objects can be logged with the failure integration. This works really well if you use the failure::Error type or if you have failure::Fail objects that use the failure context internally to gain a backtrace.

Example

use sentry_failure::capture_error;
let result = match function_that_might_fail() {
    Ok(result) => result,
    Err(err) => {
        capture_error(&err);
        return Err(err);
    }
};

To capture fails and not errors use capture_fail.

Structs

FailureIntegration
The Sentry Failure Integration.

Traits

FailureHubExt
Hub extension methods for working with failure.

FailureResultExt
Extension trait providing methods to unwrap a result, preserving backtraces from the underlying error in the event of a panic.

Functions

capture_error
Captures a boxed failure (failure::Error).

capture_fail
Captures a failure::Fail.

event_from_error
Helper function to create an event from a failure::Error.

event_from_fail
Helper function to create an event from a failure::Fail.

exception_from_single_fail
This converts a single fail instance into an exception.

panic_extractor
Extracts a Sentry Event from a PanicInfo.